Correct code fix logic for legacy provider
