--- src/Bicep.Core/TypeSystem/Providers/LegacyProviderSpecification.cs
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-using Bicep.Core.Parsing;
-
-namespace Bicep.Core.TypeSystem.Providers;
-
-public record LegacyProviderSpecification(string NamespaceIdentifier, string Version, bool IsValid, TextSpan Span) : IProviderSpecification;
