+++ src/Bicep.Core/TypeSystem/Providers/ProviderSpecificationFactory.cs
@@ -44,7 +44,7 @@ public static partial class ProviderSpecificationFactory
                 legacyMatch.Groups["name"].Value,
                 legacyMatch.Groups["version"].Value,
                 IsValid: true,
-                stringSyntax.GetInnerSpan());
+                stringSyntax.Span);
         }
 
         var registryMatch = InlinedSpecificationPattern().Match(value);
@@ -57,12 +57,11 @@ public static partial class ProviderSpecificationFactory
         var address = registryMatch.Groups["address"].Value;
         var version = registryMatch.Groups["version"].Value;
 
-        var span = stringSyntax.GetInnerSpan();
         // We normalize the artifact address to the way we represent module addresses, see https://github.com/Azure/bicep/issues/12202
         var unexpandedArtifactAddress = $"{address}:{version}";
         var name = RepositoryNamePattern().Match(address).Groups["name"].Value;
 
-        return new InlinedProviderSpecification(name, version, unexpandedArtifactAddress, IsValid: true, span);
+        return new InlinedProviderSpecification(name, version, unexpandedArtifactAddress, IsValid: true, stringSyntax.Span);
     }
 }
