--- src/Bicep.Core/TypeSystem/Providers/ProviderSpecificationTrivia.cs
@@ -1,12 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-using Bicep.Core.Parsing;
-
-namespace Bicep.Core.TypeSystem.Providers;
-public record ProviderSpecificationTrivia(TextSpan Span) : IProviderSpecification
-{
-    public string NamespaceIdentifier => LanguageConstants.ErrorName;
-    public bool IsValid => false;
-    public string? Version => null;
-};
